requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain>=0.1.16
langchain-google-genai>=0.0.9
langchain-core>=0.1.34
google-generativeai==0.3.2
httpx==0.25.2
python-multipart==0.0.6